Switch to MASON_LLVM_RELEASE 10.0.0 which has clang-tidy and clang-format

+++ test/test.cpp
@@ -18,7 +18,7 @@ TEST_CASE("PointTreeTest") {
   // Build a k-d tree from a list of points
 
   std::vector<Vector3, Eigen::aligned_allocator<Vector3>> points;
-  std::mt19937_64 randGen{42};
+  std::mt19937_64 randGen{size_t(42)};
   std::uniform_real_distribution<double> dist{-1000.0, 1000.0};
 
   // Make random points
